split file handling modals to fileActions.jsx
